refactor: centralize tagged and lowered abi transitions
