+++ crates/emit/src/expressions/staging.rs
@@ -76,21 +76,48 @@ impl Emitter<'_> {
         let staged = self
             .with_go_fn_short_circuit_suppressed(suppress, |this| this.stage_composite(expression));
 
-        if suppress
-            && !matches!(expression.unwrap_parens(), Expression::Lambda { .. })
-            && !Self::is_tagged_shape_fn_value(expression)
-            && self.classify_go_fn_value(expression).is_none()
-            && let Some(param_ty) = param_ty
-            && let syntax::types::Type::Function { return_type, .. } = param_ty.unwrap_forall()
-            && self.classify_direct_emission(return_type).is_some()
-        {
+        if suppress {
             let mut setup = staged.setup;
-            let cb_var = self.hoist_tmp_value(&mut setup, "cb", &staged.value);
-            let tagged = self.lower_arg_to_tagged(&mut setup, &cb_var, param_ty);
-            return EmittedExpression::new(setup, tagged, expression);
+            if let Some(tagged) =
+                self.try_lower_arg_to_tagged(&mut setup, expression, &staged.value, param_ty)
+            {
+                return EmittedExpression::new(setup, tagged, expression);
+            }
+            return EmittedExpression::new(setup, staged.value, expression);
         }
 
         staged
+    }
+
+    /// Adapt a lowered-ABI Lisette callback to tagged shape when the callee
+    /// (prelude generic or otherwise) declares a function param whose return
+    /// classifies for direct (lowered) emission. Returns `None` when the arg
+    /// already has tagged shape, is a lambda literal, or is a Go fn value.
+    pub(crate) fn try_lower_arg_to_tagged(
+        &mut self,
+        output: &mut String,
+        arg: &Expression,
+        value: &str,
+        param_ty: Option<&Type>,
+    ) -> Option<String> {
+        if matches!(arg.unwrap_parens(), Expression::Lambda { .. }) {
+            return None;
+        }
+        if Self::is_tagged_shape_fn_value(arg) {
+            return None;
+        }
+        if self.classify_go_fn_value(arg).is_some() {
+            return None;
+        }
+        let param_ty = param_ty?;
+        let Type::Function { return_type, .. } = param_ty.unwrap_forall() else {
+            return None;
+        };
+        self.classify_direct_emission(return_type)?;
+        let cb_var = self.hoist_tmp_value(output, "cb", value);
+        Some(crate::types::abi_transition::lower_arg_to_tagged(
+            self, output, &cb_var, param_ty,
+        ))
     }
 
     pub(crate) fn stage_native_method_args(
