+++ crates/emit/src/types/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod abi;
+pub(crate) mod abi_transition;
 pub(crate) mod coercion;
 pub(crate) mod emitter;
 pub(crate) mod go_type;
